requirements.txt
a2a-sdk[http-server]>=0.2.7
litellm>=1.60.0
uvicorn>=0.27.0
